Revert "Address review feedback: extract modules, clean up imports (#90813)"
This reverts commit c83f475c1c76c2a23dd5cadd4f92809c979d688a.
