+++ test/production/next-server-nft/next-server-nft.test.ts
@@ -137,10 +137,17 @@ async function readNormalizedNFT(next, name) {
            "/node_modules/next/dist/build/adapter/setup-node-env.external.js",
            "/node_modules/next/dist/build/define-env.js",
            "/node_modules/next/dist/build/duration-to-string.js",
+           "/node_modules/next/dist/build/get-supported-browsers.js",
            "/node_modules/next/dist/build/next-config-ts/require-hook.js",
            "/node_modules/next/dist/build/next-config-ts/transpile-config.js",
+           "/node_modules/next/dist/build/output/format.js",
            "/node_modules/next/dist/build/output/log.js",
+           "/node_modules/next/dist/build/segment-config/app/app-segment-config.js",
+           "/node_modules/next/dist/build/segment-config/app/app-segments.js",
+           "/node_modules/next/dist/build/segment-config/app/collect-root-param-keys.js",
+           "/node_modules/next/dist/build/static-paths/app.js",
            "/node_modules/next/dist/build/static-paths/app/extract-pathname-route-param-segments-from-loader-tree.js",
+           "/node_modules/next/dist/build/static-paths/pages.js",
            "/node_modules/next/dist/build/static-paths/utils.js",
            "/node_modules/next/dist/build/swc/helpers.js",
            "/node_modules/next/dist/build/swc/index.js",
@@ -149,7 +156,7 @@ async function readNormalizedNFT(next, name) {
            "/node_modules/next/dist/build/swc/loaderWorkerPool.js",
            "/node_modules/next/dist/build/swc/options.js",
            "/node_modules/next/dist/build/swc/types.js",
-           "/node_modules/next/dist/build/webpack-layer.js",
+           "/node_modules/next/dist/build/utils.js",
            "/node_modules/next/dist/cli/next-test.js",
            "/node_modules/next/dist/client/*",
            "/node_modules/next/dist/compiled/@edge-runtime/cookies/index.js",
@@ -167,6 +174,7 @@ async function readNormalizedNFT(next, name) {
            "/node_modules/next/dist/compiled/@napi-rs/triples/index.js",
            "/node_modules/next/dist/compiled/@opentelemetry/api/index.js",
            "/node_modules/next/dist/compiled/async-retry/index.js",
+           "/node_modules/next/dist/compiled/async-sema/index.js",
            "/node_modules/next/dist/compiled/busboy/index.js",
            "/node_modules/next/dist/compiled/bytes/index.js",
            "/node_modules/next/dist/compiled/ci-info/index.js",
@@ -210,6 +218,7 @@ async function readNormalizedNFT(next, name) {
            "/node_modules/next/dist/compiled/strip-ansi/index.js",
            "/node_modules/next/dist/compiled/superstruct/index.cjs",
            "/node_modules/next/dist/compiled/tar/index.js",
+           "/node_modules/next/dist/compiled/text-table/index.js",
            "/node_modules/next/dist/compiled/watchpack/watchpack.js",
            "/node_modules/next/dist/compiled/ws/index.js",
            "/node_modules/next/dist/compiled/zod-validation-error/index.js",
@@ -219,6 +228,7 @@ async function readNormalizedNFT(next, name) {
            "/node_modules/next/dist/experimental/testmode/httpget.js",
            "/node_modules/next/dist/experimental/testmode/server-edge.js",
            "/node_modules/next/dist/experimental/testmode/server.js",
+           "/node_modules/next/dist/export/helpers/create-incremental-cache.js",
            "/node_modules/next/dist/lib/batcher.js",
            "/node_modules/next/dist/lib/build-custom-route.js",
            "/node_modules/next/dist/lib/bundler.js",
